# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bissi / Committee / Chit Fund Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List tokens
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: committeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
    post:
      operationId: createToken
      tags: [tokens]
      summary: Allocate a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenInput"
      responses:
        "201":
          description: Token allocated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
  /tokens/{id}:
    get:
      operationId: getToken
      tags: [tokens]
      summary: Get token by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
    patch:
      operationId: updateToken
      tags: [tokens]
      summary: Update a token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenUpdate"
      responses:
        "200":
          description: Token updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"

  # Loans
...
  /reports/collection:
    get:
      operationId: getCollectionReport
      tags: [reports]
      summary: Get collection report
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
        - name: collectorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Collection report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionReport"
  /reports/loan:
    get:
      operationId: getLoanReport
      tags: [reports]
      summary: Get loan report
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Loan report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanReport"
  /reports/customer-statement:
    get:
      operationId: fetchCustomerStatement
      tags: [reports]
      summary: Get customer account statement
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customer statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerStatement"
components:
  schemas:
...
    TrendPoint:
      type: object
      required: [date, amount, count]
      properties:
        date:
          type: string
        amount:
          type: number
        count:
          type: integer
...
    Customer:
      type: object
      required: [id, referenceNumber, name, mobile, status, branchId, createdAt]
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        name:
          type: string
        mobile:
          type: string
        alternateMobile:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        aadhaar:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        nomineeNname:
          type: ["string", "null"]
        nomineeRelation:
          type: ["string", "null"]
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blocked]
        totalTokens:
          type: integer
        totalLoans:
          type: integer
        totalPaid:
          type: number
        createdAt:
          type: string
...
    PassbookEntry:
      type: object
      required: [id, type, description, amount, date]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [payment, loan_disbursed, loan_emi, lottery_win, token_allocation]
        description:
          type: string
        amount:
          type: number
        balance:
          type: ["number", "null"]
        date:
          type: string
...
    Token:
      type: object
      required: [id, tokenNumber, customerId, committeeId, status, createdAt]
      properties:
        id:
          type: integer
        tokenNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        committeeId:
          type: integer
        committeeName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, transferred, closed]
        createdAt:
          type: string
    TokenInput:
      type: object
      required: [customerId, committeeId]
      properties:
        customerId:
          type: integer
        committeeId:
          type: integer
        tokenNumber:
          type: string
    TokenUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, transferred, closed]
        transferToCustomerId:
          type: integer
...
    CollectionReport:
      type: object
      required: [startDate, endDate, totalAmount, totalCount, byBranch, byPaymentMode, daily]
      properties:
        startDate:
          type: string
        endDate:
          type: string
        totalAmount:
          type: number
        totalCount:
          type: integer
        byBranch:
          type: array
          items:
            $ref: "#/components/schemas/ReportBranchRow"
        byPaymentMode:
          type: array
          items:
            $ref: "#/components/schemas/ReportModeRow"
        daily:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
    ReportBranchRow:
      type: object
      required: [branchName, amount, count]
      properties:
        branchName:
          type: string
        amount:
          type: number
        count:
          type: integer
    ReportModeRow:
      type: object
      required: [mode, amount, count]
      properties:
        mode:
          type: string
        amount:
          type: number
        count:
          type: integer
    LoanReport:
      type: object
      required: [startDate, endDate, totalLoans, totalDisbursed, totalCollected, totalOverdue]
      properties:
        startDate:
          type: string
        endDate:
          type: string
        totalLoans:
          type: integer
        totalDisbursed:
          type: number
        totalCollected:
          type: number
        totalOverdue:
          type: number
        byBranch:
          type: array
          items:
            $ref: "#/components/schemas/ReportBranchRow"
    CustomerStatement:
      type: object
      required: [customer, startDate, endDate, entries, openingBalance, closingBalance]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        startDate:
          type: string
        endDate:
          type: string
        openingBalance:
          type: number
        closingBalance:
          type: number
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PassbookEntry"